+++ src/helpers/index.ts
@@ -37,3 +37,4 @@ export * from './random-id';
 export * from './throttle';
 export * from './shadow-dom-utils';
 export * from './node-text-utils';
+export * from './value-matchers';