+++ tests/scriptlets/index.test.js
@@ -55,3 +55,4 @@ import './inject-css-in-shadow-dom.test';
 import './remove-node-text.test';
 import './trusted-replace-node-text.test';
 import './trusted-prune-inbound-object.test';
+import './trusted-suppress-native-method.test';