+++ CHANGELOG.md
@@ -11,12 +11,13 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 
 ### Added
 
+- `trusted-suppress-native-method` scriptlet [#383]
 - `json-prune-fetch-response` scriptlet [#361]
 - `json-prune-xhr-response` scriptlet [#360]
 - `href-sanitizer` scriptlet [#327]
 - `no-protected-audience` scriptlet [#395]
 - Domain value for setting cookie scriptlets [#389]
-- Multiple redirects can be used as scriptlets [#300]:
+- Multiple redirects can now be used as scriptlets [#300]:
     - `amazon-apstag`
     - `didomi-loader`
     - `fingerprintjs2`
@@ -53,6 +54,7 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 [#395]: https://github.com/AdguardTeam/Scriptlets/issues/395
 [#389]: https://github.com/AdguardTeam/Scriptlets/issues/389
 [#388]: https://github.com/AdguardTeam/Scriptlets/issues/388
+[#383]: https://github.com/AdguardTeam/Scriptlets/issues/383
 [#377]: https://github.com/AdguardTeam/Scriptlets/issues/377
 [#361]: https://github.com/AdguardTeam/Scriptlets/issues/361
 [#360]: https://github.com/AdguardTeam/Scriptlets/issues/360